refactor: create seperate class for donation, donor and form
